NgocNTH: fix form buttons layout
